Added Vertically Moving Platforms
Once again have not had time to playtest, it is likely they player will
just fall through them or something but they should render and move
correctly.

Mostly copy/paste from regular moving platforms with any references to
the x-axis changed to the y-axis

Added in a min and max Y value, this is to prevent me from needing to
add a roof to any level and to avoid player being collided between
platform and roof.

Added method for collision checks to make adding a check for the min and
max bounds easier